Split simple/sticky OnThisPageAnchors into their own variant files
